requirements.txt
gradio>=4.0.0
gradio_client
pandas
requests
pdfplumber>=0.10.0
Pillow>=10.0.0
transformers
torch
huggingface_hub
datasets
